requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.5
defusedxml==0.5.0
Django==2.1.5
django-el-pagination==3.2.4
django-smuggler==0.9.1
djangorestframework==3.9.1
httplib2==0.12.0
idna==2.8
lxml==4.3.0
mysqlclient==1.4.1
oauth2==1.9.0.post1
oauthlib==3.0.1
Pillow==5.4.1
pycparser==2.19
PyJWT==1.7.1
python-openid==2.2.5
python3-openid==3.1.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.0.0
soupsieve==1.7.3
urllib3==1.24.1
